Feature/bid

+++ final-project-springboots/src/main/java/com/springbootstudy/bbs/domain/GradeVO.java
@@ -2,12 +2,12 @@ package com.springbootstudy.bbs.domain;
 
 import lombok.Data;
 
-//회원 등급 코드 테이블
+// 회원 등급 코드 테이블
 @Data
-public class GradeVO { 
+public class GradeVO {
 
- private Integer gradeIdx;     // PK
- private String  gradeName;    // 등급명 (basic / silver / gold / vip)
- private Double  gradeCredit;  // 신용도 기준 
+    private Integer gradeIdx;     // PK
+    private String  gradeName;    // 등급명 (basic / silver / gold / vip)
+    private Integer gradeCredit;  // 등급 기준 크레딧 (예: normal 0, bronze 500, silver 2000, gold 5000, vip 10000)
 
 }
